接入第三方库实现音频暂停播放及转码为mp3

+++ app/src/main/java/com/example/snowson/apptest/utils/AudioFileUtils.java
@@ -128,6 +128,9 @@ public class AudioFileUtils {
         File rootFile = new File(fileBasePath);
         if (rootFile.exists()) {
             for (File item : rootFile.listFiles()) {
+                if (item.getName().equals("result.pcm")) {
+                    continue;
+                }
                 item.delete();
             }
         }
